выолнил последнюю задачу в файле check_google.py

+++ check_google.py
@@ -1,27 +1,21 @@
-import subprocess
+import socket
 import os
 
 servers = ['drive.google.com', 'mail.google.com', 'google.com']
-
+if os.path.exists('out.txt'):
+    print('file exist')
+    with open('out.txt', 'r') as f:
+        for list_file in f:
+            k, v = list_file.strip().split()
+            ip = socket.gethostbyname(k)
+            if ip != v:
+                print('[ERROR]', k,  'IP mismatch:', v, '=>', ip)
+    os.remove('out.txt')
+else:
+    print('file not exist, nothing to compare')
 
 for server in servers:
-    ping = subprocess.Popen(["ping", "-n", "1", server], stdout = subprocess.PIPE, stderr = subprocess.PIPE)
-    data = str(ping.communicate())
-    list = []
-    for line in data.split():
-        list.append(line)
-    formatted_list = list[1:3]
-#    print(formatted_list)
+    ip = socket.gethostbyname(server)
+    print(server, ip)
     with open('out.txt', 'a') as f:
-        print(list[1:3], file=f)
-#    print(list[1:3])
-with open('out.txt', 'r') as file:
-    for i in file:
-        for k in formatted_list:
-            if i == k:
-                print(equal)
-            else:
-                print('[ERROR] <URL сервиса> IP mismatch: <старый IP> <Новый IP>')
-os.remove('out.txt')
-
-
+        print(server, ip, file=f)
